[REFACTOR]: move utils used for tests out of include
