+++ lib/CMakeLists.txt
@@ -6,7 +6,6 @@ include_directories(include private)
 
 add_library (
     ConcurrencyLib
-    "src/table.cpp"
     "src/thread_pool.cpp"
 )
