+++ tests/contract_test.cpp
@@ -5,8 +5,8 @@
 #include <thread>
 
 #include "utils/logger.hpp"
-#include "utils/timer.hpp"
-#include "utils/tester.hpp"
+#include "test_utils/timer.hpp"
+#include "test_utils/tester.hpp"
 
 #include "contract.hpp"
