+++ tests/linear_equations_test.cpp
@@ -1,10 +1,10 @@
 #include <cmath>
 
 #include "utils/logger.hpp"
-#include "utils/timer.hpp"
-#include "utils/tester.hpp"
-#include "utils/matrix.hpp"
-#include "utils/linalg.hpp"
+#include "test_utils/timer.hpp"
+#include "test_utils/tester.hpp"
+#include "test_utils/matrix.hpp"
+#include "test_utils/linalg.hpp"
 
 #include "async_function.hpp"
 #include "thread_pool.hpp"
