+++ tests/thread_pool_test.cpp
@@ -12,8 +12,8 @@
 #include <vector>
 
 #include "utils/logger.hpp"
-#include "utils/timer.hpp"
-#include "utils/tester.hpp"
+#include "test_utils/timer.hpp"
+#include "test_utils/tester.hpp"
 
 #include "thread_pool.hpp"
 #include "async_function.hpp"
